refactor: narrow runtime provider surfaces
