+++ typescript/widgets/src/walletIntegrations/AccountList.tsx
@@ -1,7 +1,7 @@
 import { clsx } from 'clsx';
 import React, { ButtonHTMLAttributes } from 'react';
 
-import type { MultiProtocolProvider } from '@hyperlane-xyz/sdk/providers/MultiProtocolProvider';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '@hyperlane-xyz/sdk/providers/ConfiguredMultiProtocolProvider';
 import type { ChainName } from '@hyperlane-xyz/sdk/types';
 import { ProtocolType, objKeys } from '@hyperlane-xyz/utils';
