+++ typescript/widgets/src/walletIntegrations/accounts.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 
-import type { MultiProtocolProvider } from '@hyperlane-xyz/sdk/providers/MultiProtocolProvider';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '@hyperlane-xyz/sdk/providers/ConfiguredMultiProtocolProvider';
 import type { ChainName } from '@hyperlane-xyz/sdk/types';
 import { type Address, type KnownProtocolType, ProtocolType } from '@hyperlane-xyz/utils';
