+++ typescript/widgets/src/walletIntegrations/transactionFns.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 
-import type { MultiProtocolProvider } from '@hyperlane-xyz/sdk/providers/MultiProtocolProvider';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '@hyperlane-xyz/sdk/providers/ConfiguredMultiProtocolProvider';
 import { type KnownProtocolType, ProtocolType } from '@hyperlane-xyz/utils';
 
 import { useAleoTransactionFns } from './aleo.js';
